Split the redux types
