+++ packages/nexusgraph-provider/index.ts
@@ -1,4 +1,6 @@
 // Copyright 2023 Paion Data. All rights reserved.
 export { default as StoreWrapper } from "./src/StoreProvider";
 export type { GlobalState } from "./src/globalState";
-export * from "./src/types";
+export * from "./src/astraios/astraiosTypes";
+export * from "./src/editor/editorTypes";
+export * from "./src/nlp/nlpTypes";
