+++ packages/nexusgraph-provider/src/globalState.ts
@@ -1,5 +1,6 @@
 // Copyright 2023 Paion Data. All rights reserved.
-import { NLP_DATA as nlpData, NlpState, EDITOR_STATE as editor } from "./types";
+import { NLP_DATA as nlpData, NlpState } from "./nlp/nlpTypes";
+import { EDITOR_STATE as editor } from "./editor/editorTypes";
 import { AstraiosState, NOTE_STATE as astraios } from "./astraios/astraiosTypes";
 
 export interface GlobalState {
