+++ packages/nexusgraph-provider/src/nlp/nlpTypes.ts
@@ -22,11 +22,3 @@ export interface NlpAction {
   type: typeof UPDATE_NLPDATA;
   payload: NlpState;
 }
-
-export const EDITOR_STATE = "editor";
-export const UPDATE_EDITOR_STATE = EDITOR_STATE + "/UPDATE_EDITOR_STATE";
-
-export interface EditorAction {
-  type: typeof UPDATE_EDITOR_STATE;
-  payload: object;
-}
