+++ packages/nexusgraph-provider/src/rootReducer.ts
@@ -1,8 +1,9 @@
 import nlpReducer from "./nlp/nlpDuck";
+import { NLP_DATA as nlpData } from "./nlp/nlpTypes";
 import editorReducer from "./editor/editorDuck";
-import { NLP_DATA as nlpData, EDITOR_STATE as editor } from "./types";
-import { NOTE_STATE as astraios } from "./astraios/astraiosTypes";
+import { EDITOR_STATE as editor } from "./editor/editorTypes";
 import astraiosReducer from "./astraios/astraiosDuck";
+import { NOTE_STATE as astraios } from "./astraios/astraiosTypes";
 
 export default {
   [nlpData]: nlpReducer,
